GAT/requirements.txt
pandas
torch_geometric
torch_scatter
tqdm
torch_sparse
